Game over functionality was added

+++ src/Engine.js
@@ -11,9 +11,13 @@ export class Engine {
 		this.zombies = this.create2DArray();
 		this.plants = this.create2DArray();
 		this.peas = this.create2DArray();
+		this.event = new Event();
 		this.menu = menu;
 		this.sunflowerCount = 0;
-		this.fieldWidth;
+		this.timerID;
+		this.isStopped = false;
+		this.isZombieGenerated = false;
+		this.isSunPointsGenerated = false;
 	}
 
 	async createPlant(plantData) {
@@ -29,74 +33,79 @@ export class Engine {
 		this.plants[plantData.rowIndex].push(plant);
 
 		if ( plant instanceof Peashooter) {
-			await Utils.pause(2000);
+			await Utils.pause(2000);			///////////////   in constants ///  edit title img
 
 			this.shootByPea(plant, plantData.rowIndex);
 		}
 
 		if (plant instanceof Sunflower) {
-			if (this.sunflowerCount == 0) {
+			if (!this.isSunPointsGenerated) {
 				this.generateSunPoints();
+				this.isSunPointsGenerated = true;
 			}
 			this.sunflowerCount++;
 		}
 	};
 
 	async createZombie(containers) {
-		let type = ZOMBIE_DATA.TYPE[random(1)];
-		let index = random(containers.length - 1);
-		let container = containers[index];
+		if (!this.isStopped) {
+			let type = ZOMBIE_DATA.TYPE[random(1)];
+			let index = random(containers.length - 1);
+			let container = containers[index];
 
-		let zombie = new type(ENTITY_DATA.MAX_HEALTH, container);
-		
-		let event = new Event();
-		event.onKilled(zombie.die.bind(zombie));
-		event.onDeleted(this.removeInstanceFromArr.bind(this, this.zombies[index], zombie));
-		event.onDeleted(this.deleteInstance.bind(this, zombie));
-		zombie.event = event;
+			let zombie = new type(ENTITY_DATA.MAX_HEALTH, container);
+			
+			let event = new Event();
+			event.onKilled(zombie.die.bind(zombie));
+			event.onDeleted(this.removeInstanceFromArr.bind(this, this.zombies[index], zombie));
+			event.onDeleted(this.deleteInstance.bind(this, zombie)); 
+			zombie.event = event;
 
-		zombie.create();
-		this.zombies[index].push(zombie);
+			zombie.create();
+			this.zombies[index].push(zombie);
+		}
 
 		await Utils.pause(ZOMBIE_DATA.CREATE_TIMEOUT);
-
 		this.createZombie(containers);
 	}
 
 	async generateSunPoints() {
-		this.menu.setSunPoints(this.sunflowerCount * PLANT_DATA.SUN_POINTS);
-
+		if (!this.isStopped) {
+			this.menu.setSunPoints(this.sunflowerCount * PLANT_DATA.SUN_POINTS);
+		}
 		await Utils.pause(PLANT_DATA.GENERATE_POINTS_TIMEOUT);
 		this.generateSunPoints();
 	}
 
 	async shootByPea(peashooter, rowIndex) {
-		let pea = peashooter.shoot();
-		let event = new Event();
-		event.onDeleted(this.removeInstanceFromArr.bind(this, this.peas[rowIndex], pea));
-		event.onDeleted(this.deleteInstance.bind(this, pea));
-		pea.event = event;
+		if (!this.isStopped) {
+			let pea = peashooter.shoot();
+			let event = new Event();
 
-		this.peas[rowIndex].push(pea);
+			event.onDeleted(this.removeInstanceFromArr.bind(this, this.peas[rowIndex], pea));
+			event.onDeleted(this.deleteInstance.bind(this, pea));
+			pea.event = event;
+
+			this.peas[rowIndex].push(pea);
+		}
 
 		await Utils.pause(PLANT_DATA.SHOOT_TIMEOUT);
-
 		if (peashooter.health > 0) {
 			this.shootByPea(peashooter, rowIndex);
 		}
 	};
 
 	create2DArray() {
-		var arr = [];
+		let arr = [];
 
-		for (var i = 0; i < SETTINGS.ROW_COUNT; i++) {
+		for (let i = 0; i < SETTINGS.ROW_COUNT; i++) {
 			arr[i] = [];
 		}
 
 		return arr;
 	}
 
-	animate() {
+	animate(fieldWidth) {
 		const frame = () => {
 			that.zombies.forEach((row, rowIndex) => {
 				
@@ -113,7 +122,7 @@ export class Engine {
 					}
 
 					if (zombie.position <= 0) {
-						clearInterval(id);
+						Utils.triggerEvent(this.event.onGameOverEvt);
 					}
 				});
 			});
@@ -129,7 +138,7 @@ export class Engine {
 							pea.delete();
 						}
 					}
-					if (pea.position >= this.fieldWidth - pea.width) {
+					if (pea.position >= fieldWidth - pea.width) {
 						pea.delete();
 					} else {
 						pea.move();
@@ -139,23 +148,24 @@ export class Engine {
 		}
 
 		let that = this;
-		let id = setInterval(frame, SETTINGS.INTERVAL);
+		this.timerID = setInterval(frame, SETTINGS.INTERVAL);
 	}
 	
 	async hitPlant(plant, zombie) {
-		plant.hit(ENTITY_DATA.HIT_DAMAGE);
+		if (!this.isStopped) {
+			plant.hit(ENTITY_DATA.HIT_DAMAGE);
+
+			if (plant.health <= 0 ) {
+				if (plant instanceof Sunflower) {
+					this.sunflowerCount--;
+				}
+				plant.kill();
+			}
+		}
 
 		if (plant.health > 0 && zombie.health > 0) {
 			await Utils.pause(PLANT_DATA.DAMAGE_TIMEOUT);
 			this.hitPlant(plant, zombie);
-		}
-
-		if (plant.health <= 0 ) {
-			plant.kill();
-
-			if (plant instanceof Sunflower) {
-				this.sunflowerCount--;
-			}
 		}
 	}
 
@@ -199,5 +209,24 @@ export class Engine {
 
 	removeInstanceFromArr(arr, instance) {
 		pull(arr, instance);
+	}
+
+	deleteAllEntities() {
+		this.clear2DArray(this.zombies);
+		this.clear2DArray(this.plants);
+		this.clear2DArray(this.peas);
+
+		this.sunflowerCount = 0;
+	}
+
+	clear2DArray(entity2DArray) {
+		for (let i = 0; i < entity2DArray.length; i++) {
+			let rowLength = entity2DArray[i].length;
+
+			while (rowLength > 0) {
+				entity2DArray[i][0].delete();
+				rowLength--;
+			}
+		}
 	}
 }
